fix(plugin-sdk): keep setup adapter contract acyclic
