+++ docs/.generated/plugin-sdk-api-baseline.sha256
@@ -43,7 +43,7 @@ bba5540be7cf9613a163663decdb2affe2af9bbd3ad7914989ab186f9c2abec1  module/channel
 17cec26bc71fc43a066049ef63f95bf29737113c26ab13689ceff602b9aa11d6  module/channel-secret-basic-runtime
 0ceb4378709eb2d92a62a275f87fa04e18f77df8f942a9a0acef81019ebc1e24  module/channel-secret-runtime
 7c90157a95bc0523fc66b1f78ce140f7ec7dbf809dfa3a480244efc01f972754  module/channel-send-result
-fa58237f4af17c2a3444eb01475e7267d86f37ffc8bab6ce243143a35908bbfd  module/channel-setup
+fb123c1b557ed2527e335f13c3d6de41ab0c3305151001cf2b1075d1da8034c5  module/channel-setup
 47719805c7a1d2623dcecafad50abb6cc674262b250e1f0d7020a79ad41cfd42  module/channel-status
 824858dccff862ba6b83f6a478bea73397972a4584f4a427a794224ac680670a  module/channel-streaming
 c89ec1b194b76f67a6f4dd108dccf460da6065646cba31374c8aa748f23a39e4  module/collection-runtime
@@ -119,8 +119,8 @@ b6b8edc50ecab8386c9acd8f374a207212b5a99c8f518538bbcf0c458dda3881  module/runtime
 596a315d426121c9620b314e3a9a7f523840b46e007d94d0d5e83cdedf789d15  module/security-runtime
 31b785e74f1f8f56241b7756ef6a5d86199c5ce177cbb1c234a261866972f270  module/session-discussion
 32fb6d253abf22440bc76c7a68d1f35fc0ef369b0ad738aedba9c3054a76e48e  module/session-store-runtime
-7cad408673562b0ff33f60071698bea2c39d02c64f4f235851c5c2ceb946910d  module/setup
-fbcb853789db5a1ad5d81e8094ed7f5bcea657ac2281e7887a19c1d6f1194571  module/setup-runtime
+23cc02cbfb0a0bfa41adc8f02f5f22738781495250c8cee06c18f18bfd283afe  module/setup
+21071e8c2ef020685aba09b5661e37e7415938ff6602053fa831ac9d58673248  module/setup-runtime
 cd431f6ba8327b81438b7a63b1963120f200f5abd145fb6aa7c5c561339cb0b1  module/setup-tools
 18e384ec43d9eaee52c8e286e127bda2048370e2337964a754d94b236724ca9e  module/skill-commands-runtime
 ae469f32799380e6b045abaefefee6eb3f00d714ffbf36b6eeef5025dc529472  module/speech-settings
