+++ src/channels/plugins/setup-adapter.types.ts
@@ -1,0 +1,45 @@
+import type { OpenClawConfig } from "../../config/types.openclaw.js";
+import type { RuntimeEnv } from "../../runtime.js";
+import type { ChannelSetupInput } from "./setup-input.js";
+
+export type ChannelSetupAdapter<Input extends { name?: string } = ChannelSetupInput> = {
+  resolveAccountId?: (params: { cfg: OpenClawConfig; accountId?: string; input?: Input }) => string;
+  prepareAccountConfigInput?: (params: {
+    cfg: OpenClawConfig;
+    accountId: string;
+    input: Input;
+    runtime: RuntimeEnv;
+  }) => Promise<Input> | Input;
+  resolveBindingAccountId?: (params: {
+    cfg: OpenClawConfig;
+    agentId: string;
+    accountId?: string;
+  }) => string | undefined;
+  applyAccountName?: (params: {
+    cfg: OpenClawConfig;
+    accountId: string;
+    name?: string;
+  }) => OpenClawConfig;
+  applyAccountConfig: (params: {
+    cfg: OpenClawConfig;
+    accountId: string;
+    input: Input;
+  }) => OpenClawConfig;
+  afterAccountConfigWritten?: (params: {
+    previousCfg: OpenClawConfig;
+    cfg: OpenClawConfig;
+    accountId: string;
+    input: Input;
+    runtime: RuntimeEnv;
+  }) => Promise<void> | void;
+  validateInput?: (params: {
+    cfg: OpenClawConfig;
+    accountId: string;
+    input: Input;
+  }) => string | null;
+  singleAccountKeysToMove?: readonly string[];
+  namedAccountPromotionKeys?: readonly string[];
+  resolveSingleAccountPromotionTarget?: (params: {
+    channel: Record<string, unknown>;
+  }) => string | undefined;
+};
