+++ src/channels/plugins/setup-contract.ts
@@ -3,8 +3,8 @@ import { Option } from "commander";
 import type { OpenClawConfig } from "../../config/types.openclaw.js";
 import { parseStrictNonNegativeInteger } from "../../infra/parse-finite-number.js";
 import type { RuntimeEnv } from "../../runtime.js";
+import type { ChannelSetupAdapter } from "./setup-adapter.types.js";
 import type { ChannelSetupInput } from "./setup-input.js";
-import type { ChannelSetupAdapter } from "./types.adapters.js";
 
 type ChannelSetupCliOption = {
   flags: string;
